requirements.txt
nuitka  == 0.5.11
PyMonad == 1.3
PyYAML  == 3.11
nose    == 1.3.4
rednose == 0.4.1
